Add Example 1.9.c
Example 1.9 is a program which reads in an input and then prints the
longest line. It uses a loop which calls a function named getline() to
read the input line-by-line. If we think that the most recent line is
the longest so far, then we save it away via a function named copy().
Whichever line we think is the longest at the end of the loop, we print
out.