Update Exercise 2-6.c
Changed so that setbits() sets the n bits to the right of p (inclusive)
as opposed to the n bits to the left (inclusive). The brief doesn't make
this clear, but online there seems to be more examples of people that
interpreted it this way rather than the other way.