Add Example 4.3.c
Example 4.3 is the implementation of a "reverse Polish" calculator.
According to the book, the following calculation...

(1 - 2) * (4 + 5)

...can be written in reverse Polish as follows:

1 2 - 4 5 + *

The calculator operates on 1 and 2 and replaces them with the result, -1
(e.g. 1 - 2). The calculator then operates on 4 and 5 and replaces them with
the result, 9 (e.g. 4 + 5). This leave us with the following expression

-1 9 *

The calculator performs this operation and ends up with the result, -9.

The example is taken from page 76.